Went back to dictionary-based header.
After looking at some FITS files, I realized that keywords appear to be
arbirary, not just limited to those defined by the standard, so I am
going back to defining a few that seem to be more or less required, and
plugging the rest into a dictionary-like object by type. We'll see how
this works.