Refactored data element, updated notebook.
Refactored the data element to just be a list of each possible numeric
type. This simplifies the access/setting syntax and seems like it should
work since we have information about the size of any data incorporated
in the header. Updated the notebook to grab headers/data from a sample
FITS file and plug them into the protobuf.